// hackathon/react/my-app-ts/src/types/User.ts
import React from 'react'

export type UserName = {
    name: string
};

export type userPost = {
    userId: string
    name: string
    affiliationId: string
    points: number
};

export type UserOption = {
    // value1: string;
    value: string;
    label: string;
    // affiliationId: string;
    // points: number;
};